+++ scripts/gen_doc.sh
@@ -1,2 +1,2 @@
-ai --raw translate -f README_zh.md -t en >README.md
-ai --raw translate -f README_zh.md -t ja >README_ja.md
+ai --raw -b GPTBot translate -f README_zh.md -t en >README.md
+ai --raw -b GPTBot translate -f README_zh.md -t ja >README_ja.md